backend/requirements.txt
amqp==5.3.1
asgiref==3.7.2
attrs==25.4.0
autobahn==25.11.1
Automat==25.4.16
billiard==4.2.4
cbor2==5.7.1
celery==5.4.0
certifi==2025.11.12
cffi==2.0.0
channels==4.0.0
channels-redis==4.0.0
charset-normalizer==3.4.4
click==8.3.1
click-didyoumean==0.3.1
click-plugins==1.1.1.2
click-repl==0.3.0
colorama==0.4.6
constantly==23.10.4
cron_descriptor==2.0.6
cryptography==46.0.3
daphne==4.2.1
defusedxml==0.7.1
Django==4.2
django-celery-beat==2.8.1
django-cors-headers==4.0.0
django-templated-mail==1.1.1
django-timezone-field==7.2.1
django_celery_results==2.6.0
djangorestframework==3.14.0
djangorestframework_simplejwt==5.5.1
djoser==2.2.0
drf-spectacular==0.29.0
drf-yasg==1.21.11
gunicorn==20.1.0
hyperlink==21.0.0
idna==3.11
Incremental==24.11.0
inflection==0.5.1
jsonschema==4.25.1
jsonschema-specifications==2025.9.1
kombu==5.6.1
msgpack==1.1.2
oauthlib==3.3.1
packaging==25.0
prompt_toolkit==3.0.52
psycopg2-binary==2.9.11
py-ubjson==0.16.1
pyasn1==0.6.1
pyasn1_modules==0.4.2
pycparser==2.23
PyJWT==2.8.0
pyOpenSSL==25.3.0
python-crontab==3.3.0
python-dateutil==2.9.0.post0
python-dotenv==1.0.0
python3-openid==3.2.0
pytz==2025.2
PyYAML==6.0.3
redis==4.5.5
referencing==0.37.0
requests==2.32.5
requests-oauthlib==2.0.0
rpds-py==0.30.0
service-identity==24.2.0
setuptools==80.9.0
six==1.17.0
social-auth-app-django==5.4.3
social-auth-core==4.7.0
sqlparse==0.5.4
Twisted==25.5.0
txaio==25.12.1
typing_extensions==4.15.0
tzdata==2025.2
ujson==5.11.0
uritemplate==4.2.0
urllib3==2.5.0
vine==5.1.0
wcwidth==0.2.14
wheel==0.45.1
zope.interface==8.1.1
